+++ src/components/DetailTopic/styled.js
@@ -44,25 +44,44 @@ export const MainContainer = styled.div`
     margin-top: 10px;
   }
 
-  textarea {
-    overflow: hidden;
-    padding: 10px;
-    width: 99%;
-    height: 150px;
-    font-size: 14px;
-    display: block;
-    margin-bottom: 15px;
-  }
-
-  button {
-    margin-top: 5px;
-    width: 98%;
-    border-radius: 10px;
-    background-color: #3e2776;
-    color: white;
-  }
-
   .midButton {
     width: 24.5% !important;
   }
+`;
+
+export const Button = styled.button`
+  margin-top: 5px;
+  width: 98%;
+  border-radius: 10px;
+  background-color: #3e2776;
+  color: white;
+`;
+
+export const Comment = styled.textarea`
+  border: 1.5px solid grey;
+  overflow-y: scroll;
+  resize: none;
+  padding: 10px;
+  width: 99%;
+  height: 300px;
+  font-size: 14px;
+  display: block;
+  margin-bottom: 15px;
+  border-bottom-left-radius: 5px;
+`;
+
+export const Options = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export const OptionButton = styled.button`
+  background-color: #a9b3b8;
+  width: 30px;
+  height: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  cursor: pointer;
 `;